Refactor for extensibility.
